refactor: api calls isolated
